+++ CHANGELOG.md
@@ -7,4 +7,5 @@ All notable changes to this project will be documented in this file.
 - Config over Code implementation
 - Functionality for checking if read and write databases are in sync
 - read() method in Redis handler
+- host, port and default_db for Redis handler maintainable via config api
 - Handler for elasticsearch